Move support methods into pepys, so we can access them from Notebooks
