+++ tests/test_replay_load.py
@@ -4,8 +4,7 @@ from unittest import TestCase
 from sqlalchemy import inspect, MetaData, Table
 
 
-from .support_methods import Support_Methods
-
+from pepys_import.core.debug.support_methods import Support_Methods
 from pepys_import.core.store.data_store import DataStore
 from pepys_import.core.formats.repl_file import REPFile
 from pepys_import.core.store.sqlite_db import State
